refactor: 重构 raster tile source parser 解析模块
