+++ packages/layers/src/raster/models/index.ts
@@ -3,14 +3,12 @@ import RasterModel from './raster';
 export type RasterModelType =
   | 'raster'
   | 'raster3d'
-  | 'rasterTile'
-  | 'rasterTiff';
+  | 'rasterTile';
 
 const RasterModels: { [key in RasterModelType]: any } = {
   raster: RasterModel,
   raster3d: RasterModel,
   rasterTile: RasterTileModel,
-  rasterTiff: RasterTileModel,
 };
 
 export default RasterModels;
